requirements.txt
boto3==1.34.144
botocore==1.34.144
dnspython==2.6.1
jmespath==1.0.1
numpy==2.0.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.2
schedule==1.2.2
six==1.16.0
urllib3==1.26.19
